1/requirements.txt
asyncpg==0.30.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.0